% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_wrangling.R
\name{clean_cmd_clade}
\alias{clean_cmd_clade}
\title{Clean clade}
\usage{
clean_cmd_clade(x)
}
\arguments{
\item{x}{character. The vector of clades with CMD format.}
}
\value{
\code{clean_cmd_clade()} returns a character with
same length as \code{x}.
}
\description{
Returns only the last level name.
}
