% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_wrangling.R
\name{compute_cmd_taxtable}
\alias{compute_cmd_taxtable}
\title{Create taxonomic table}
\usage{
compute_cmd_taxtable(x)
}
\arguments{
\item{x}{character. The vector of clades with CMD format.}
}
\value{
\code{compute_cmd_taxtable()} returns a data.frame.
}
\description{
Create a taxonomic table from CMD format for clades.
}
