% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_wo_shared_zero.R
\name{cor_wo_shared_zero}
\alias{cor_wo_shared_zero}
\title{Correlation without shared zeros}
\usage{
cor_wo_shared_zero(x, y, ...)
}
\arguments{
\item{x}{a numeric vector with positive or null entries.}

\item{y}{a numeric vector with positive or null entries.}

\item{...}{arguments to be passed to \code{\link[stats]{cor}} such as \code{method} (see details).}
}
\value{
The estimated correlation coefficient between \code{x} and \code{y} after removing shared zeros.
}
\description{
Remove positions where \code{x=y=0} and then compute correlation.
}
\details{
For \code{method}: one of "pearson" (default), "kendall", or "spearman".
}
\examples{
x <- c(0, 7, 5, 0, 2)
y <- c(8, 0, 0, 0, 2)
cor_wo_shared_zero(x, y, method = "spearman")
cor(x[-4], y[-4], method = "spearman")
}
