% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_tree.R
\name{correlation_tree}
\alias{correlation_tree}
\title{Correlation tree}
\usage{
correlation_tree(table, col = 1, matrix = FALSE, remove = TRUE,
  transformation = c("reverse", "abs", "log"), ...)
}
\arguments{
\item{table}{a dataframe or a matrix.}

\item{col}{integer giving the column to be passed as row names.}

\item{matrix}{logical. Is table a matrix? Default to \code{FALSE}.}

\item{remove}{logical. If \code{TRUE} (default), shared zeros between species are removed for the computation of the pairwise correlation coefficients.}

\item{transformation}{character.}

\item{...}{arguments to be passed to \code{\link[stats]{cor}} such as \code{method} (see details).}
}
\value{
\code{correlation_tree} returns a "phylo" object.
}
\description{
Compute the correlation tree from an abundance table.
}
\details{
For \code{method}: one of "pearson" (default), "kendall", or "spearman".
}
\examples{
df <- data.frame(Clade = letters[1:4],
                 v = c(3, 0, 0, 3),
                 w = c(4, 5, 7, 0),
                 x = c(8, 5, 5, 0),
                 y = c(1, 8, 0, 0),
                 z = c(3, 5, 2, 2))
plot(correlation_tree(df, method = "spearman"))
}
