% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationtree-package.R
\docType{package}
\name{correlationtree-package}
\alias{correlationtree}
\alias{correlationtree-package}
\title{correlationtree: Functions about correlation tree}
\description{
This package provides a set of functions to
    create correlation trees and use them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abichat/correlationtree}
  \item Report bugs at \url{https://github.com/abichat/correlationtree/issues}
}

}
\author{
\strong{Maintainer}: Antoine Bichat \email{abichat@enterome.com} (0000-0001-6599-7081)

}
\keyword{internal}
