% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_cor.R
\name{cross_cor}
\alias{cross_cor}
\title{Pairwise correlations}
\usage{
cross_cor(L, remove = TRUE, .names = names(L), ...)
}
\arguments{
\item{L}{the list of all vector to be compared.}

\item{remove}{logical. If \code{TRUE} (default), shared zeros are removed.}

\item{.names}{character.}

\item{...}{arguments to be passed to \code{\link[stats]{cor}} such as \code{method} (see details).}
}
\value{
An object of class \code{dist} with all correlation between vectors.
}
\description{
Compute pairwise correlations within a list of vector.
}
\details{
For \code{method}: one of "pearson" (default), "kendall", or "spearman".
}
\examples{
df <- data.frame(Clade = letters[1:4],
                 v = c(3, 0, 0, 3),
                 w = c(4, 5, 7, 0),
                 x = c(8, 5, 5, 0),
                 y = c(1, 8, 0, 0),
                 z = c(3, 5, 2, 2))
L <- mat2list(df2mat(df))
cross_cor(L, remove = TRUE, method = "spearman")
}
