% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2mat.R
\name{df2mat}
\alias{df2mat}
\title{Dataframe to matrix}
\usage{
df2mat(df, col = 1)
}
\arguments{
\item{df}{a data frame.}

\item{col}{integer giving the column to be passed as row names.}
}
\value{
\code{df2mat} returns a named matrix.
}
\description{
Convert an abundance dataframe to a named matrix.
}
\details{
Use \code{col = 0} if there are already row names (or use \code{as.matrix(df)} directly).
}
\examples{
df <- data.frame(Clade = letters[1:4], S1 = 1:4, S2 = 7:8)
df2mat(df)
}
