% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_wrangling.R
\name{filter_cmd_level}
\alias{filter_cmd_level}
\title{Filter at sepecific level}
\usage{
filter_cmd_level(dataset, level = c("kingdom", "phylum", "class",
  "order", "family", "genus", "species", "strain"))
}
\arguments{
\item{dataset}{dataframe.}

\item{level}{characher. One of \code{c("kingdom", "phylum", "class",
"order", "family", "genus", "species", "strain")}.}
}
\description{
Filter a dataset with CMD formatting
}
