% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedist.R
\name{future_dist_BHV}
\alias{future_dist_BHV}
\title{Billera-Holmes-Vogtmann distance}
\usage{
future_dist_BHV(multiphylo)
}
\arguments{
\item{multiphylo}{a multiphylo object}
}
\value{


A distance matrix
}
\description{
Wrapper around \code{distory::dist.multiPhylo()}.
}
\details{
Could be used with
}
