% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_length.R
\name{mean_lineage_length}
\alias{mean_lineage_length}
\title{Mean lineage length}
\usage{
mean_lineage_length(tree, ultrametric = FALSE)
}
\arguments{
\item{tree}{a "phylo" object.}

\item{ultrametric}{logical. Is \code{tree} ultrametric? (see details)}
}
\value{
A double.
}
\description{
Compute the mean distance from the root to leafs.
}
\details{
A tree is ultrametric when the distances from the root to
every branch tip are equal.
}
\examples{
tree <- ape::read.tree(text = "(A:0.1,B:0.2,(C:0.3,D:0.4):0.5);")
lineage_length(tree, "A")
}
