% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedist.R
\name{pairs_grid}
\alias{pairs_grid}
\title{\code{n*(n-1)/2} unique combinations}
\usage{
pairs_grid(list)
}
\arguments{
\item{list}{list}
}
\value{
a data.frame
}
\description{
\code{n*(n-1)/2} unique combinations
}
