% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2dist.R
\name{sim2dist}
\alias{sim2dist}
\title{Convert a similarity matrix into a distance matrix}
\usage{
sim2dist(sim, transformation = "1-x")
}
\arguments{
\item{sim}{a similarity matrix of class \code{"matrix"} or \code{"dist"}.}

\item{transformation}{function or character. Transformation to apply to each element of the matrix.}
}
\value{
\code{sim2dist} returns an object of the same class as the input \code{sim}.
}
\description{
Convert a similarity matrix into a distance matrix.
}
\examples{
n <- 4
mat <- matrix(runif(n^2,-1, 1), ncol = n)
sim <- (mat + t(mat)) / 2
diag <- n * (seq_len(n) - 1) + seq_len(n)
sim[diag] <- 1
sim2dist(sim)
sim2dist(as.dist(sim), transformation = "1-absx")
}
