% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtable2tree.R
\name{taxtable2tree}
\alias{taxtable2tree}
\title{Taxonomic table to taxonomic tree}
\usage{
taxtable2tree(...)
}
\value{
error
}
\description{
Compute a taxonomic tree of class phylo from taxonomic table.
}
