% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedist.R
\name{vec2dist}
\alias{vec2dist}
\title{Transform a \code{n*(n-1)/2} vector into a distance.}
\usage{
vec2dist(vec)
}
\arguments{
\item{vec}{vector}
}
\value{
\code{dist} object
}
\description{
Transform a \code{n*(n-1)/2} vector into a distance.
}
