% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_wrangling.R
\name{compute_cmd_taxtable}
\alias{compute_cmd_taxtable}
\title{Create taxonomic table}
\usage{
compute_cmd_taxtable(...)
}
\value{
error
}
\description{
Create a taxonomic table from CMD format for clades.
}
