% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationtree-package.R
\docType{package}
\name{correlationtree-package}
\alias{correlationtree}
\alias{correlationtree-package}
\title{correlationtree: Functions about correlation tree}
\description{
This package provides a set of functions to create correlation trees and use them.
}
\author{
\strong{Maintainer}: Antoine Bichat \email{abichat@enterome.com}

}
\keyword{internal}
