% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtable2tree.R
\name{taxtable2tree}
\alias{taxtable2tree}
\title{Taxonomic table to taxonomic tree}
\usage{
taxtable2tree(df, collapse = TRUE, lineage_length = 1, root = "")
}
\arguments{
\item{df}{dataframe.}

\item{collapse}{logical. Sould node with one child be vanished?}

\item{lineage_length}{double. Mean lineage length.}

\item{root}{character. Name of the root.}
}
\value{
\code{taxtable2tree()} returns a "phylo" object.
}
\description{
Compute a taxonomic tree of class phylo from taxonomic table.
}
